[Logic] added classes: SidebarRight; Articles; Header. Remove code from class MainPage
